Made minor refactoring project
